refactor(main): extract command registration into src/commands.ts
onload registered all 19 commands inline, making the plugin entry point ~790
lines and burying lifecycle wiring (stores, view, ribbon, protocol, media
session) under a long wall of addCommand calls.

Move every command into registerCommands(plugin) in src/commands.ts; onload now
calls registerCommands(this) and drops to ~480 lines focused on lifecycle. The
two helpers the commands need (captureTimestamp, getTranscriptionService) become
public on the plugin — captureTimestamp is still also used by the media-session
handler. Command order, ids, names, icons, and guard logic are unchanged.

Behavior-preserving. Gates green; verified live (19 podnotes commands registered,
Show PodNotes routes through the module and opens the view, no console errors).
